Remove character rather than character at index

+++ WaywardCombat/src/main/java/net/wayward_realms/waywardcombat/FightImpl.java
@@ -242,7 +242,7 @@ public class FightImpl implements Fight {
     }
 
 	public void removeCharacter(Character character) {
-		turnOrder.remove(character.getId());
+		turnOrder.remove((Integer) character.getId());
 		characterLocations.remove(character.getId());
 	}
